third_party/blink/web_tests/fast/forms/input-file-set-value-expected.txt

Tests that setting the value of input type=file throws and changing type resets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.value is "example.txt"
PASS input.value is ""
PASS input.value = "foo" threw exception InvalidStateError: Failed to set the 'value' property on 'HTMLInputElement': This input element accepts a filename, which may only be programmatically set to the empty string..
PASS successfullyParsed is true

TEST COMPLETE
